chore: refactor file name from legacyUtils to fileNameUtils
