+++ cypress/e2e/resources.spec.ts
@@ -1,4 +1,4 @@
-import { slugifyCategory } from "utils/legacyUtils"
+import { slugifyCategory } from "utils/fileNameUtils"
 
 import {
   CMS_BASEURL,
